Prevent hang when there is no index source
All input source markdown files are gathered when building a site.
Internally, two golang channels are created: posts and index. The
capacity of posts was always `npages - 1`, assuming (incorrectly) that
there is always an index page. This results in a hang when building a
site with no index - the golang routine to render a page hangs because
it cannot enqueue the last page to the posts channel.

There is a second issue in that the function to render an index was also
always called. When no index is present, this function hangs on
attempting to read from the (empty) index channel.

Fix the first issue by always creating the posts channel with capacity
`npages`. Add a check to skip calling the function to render the index
if there is no index to fix the second problem.

Signed-off-by: Christopher M. Riedl <cmr@informatik.wtf>